requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
black==26.3.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
idna==3.11
iniconfig==2.3.0
isort==8.0.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.2
pytokens==0.4.1
requests==2.33.1
ruff==0.15.9
urllib3==2.6.3
